Improve Json parsing and file structure
